weasley-clock/requirements.txt
RPi.GPIO
pico
